Cleaned up chess board logic
