+++ src/client/setup/setup-base.tsx
@@ -1,6 +1,6 @@
 import { Dialog, DialogBody, DialogFooter } from "@blueprintjs/core";
 import { Outlet } from "react-router-dom";
-import { ChessboardWrapper } from "../wrapper/chessboard-wrapper";
+import { ChessboardWrapper } from "../chessboard/chessboard-wrapper";
 import { PropsWithChildren, ReactNode } from "react";
 import { Chess, DEFAULT_POSITION, Square } from "chess.js";
